requirements.txt
aiofiles==24.1.0
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.12.1
appnope==0.1.4
asttokens==3.0.1
audioop-lts==0.2.2; python_version>='3.13'
brotli==1.2.0
certifi==2026.1.4
charset-normalizer==3.4.4
click==8.3.1
coloredlogs==15.0.1
comm==0.2.3
debugpy==1.8.19
decorator==5.2.1
executing==2.2.1
fastapi==0.128.0
fastembed==0.7.4
ffmpy==1.0.0
filelock==3.20.3
flatbuffers==25.12.19
fsspec==2026.1.0
gradio==6.3.0
gradio_client==2.0.3
groovy==0.1.2
grpcio==1.76.0
h11==0.16.0
h2==4.3.0
hf-xet==1.2.0
hpack==4.1.0
httpcore==1.0.9
httpx==0.28.1
huggingface-hub==0.36.0
humanfriendly==10.0
hyperframe==6.1.0
idna==3.11
ipykernel==7.1.0
ipython==9.9.0
ipython_pygments_lexers==1.1.1
jedi==0.19.2
Jinja2==3.1.6
joblib==1.5.3
jsonpatch==1.33
jsonpointer==3.0.0
jupyter_client==8.8.0
jupyter_core==5.9.1
langchain==1.2.3
langchain_chroma
langchain-core==1.2.7
langchain_community
langchain-huggingface==1.2.0
langchain-ollama==1.0.1
langchain-qdrant==1.1.0
langchain-text-splitters==1.1.0
langgraph==1.0.5
langgraph-checkpoint==3.0.1
langgraph-prebuilt==1.0.5
langgraph-sdk==0.3.2
langsmith==0.6.2
loguru==0.7.3
markdown-it-py==4.0.0
MarkupSafe==3.0.3
matplotlib-inline==0.2.1
mdurl==0.1.2
mmh3==5.2.0
mpmath==1.3.0
nest-asyncio==1.6.0
pymupdf 
pillow
networkx==3.6.1
numpy==2.4.1
ollama==0.6.1
onnxruntime==1.23.2
orjson==3.11.5
ormsgpack==1.12.1
packaging==25.0
pandas==2.3.3
parso==0.8.5
pexpect==4.9.0
pillow==11.3.0
platformdirs==4.5.1
portalocker==3.2.0
prompt_toolkit==3.0.52
protobuf==6.33.3
psutil==7.2.1
ptyprocess==0.7.0
pure_eval==0.2.3
py_rust_stemmers==0.1.5
pydantic==2.12.5
pydantic_core==2.41.5
pydub==0.25.1
Pygments==2.19.2
PyMuPDF==1.26.6
pymupdf-layout==1.26.6
pymupdf4llm==0.2.9
python-dateutil==2.9.0.post0
python-multipart==0.0.21
pytz==2025.2
PyYAML==6.0.3
pyzmq==27.1.0
qdrant-client==1.16.2
regex==2025.11.3
requests==2.32.5
requests-toolbelt==1.0.0
rich==14.2.0
safehttpx==0.1.7
safetensors==0.7.0
scikit-learn==1.8.0
scipy==1.17.0
semantic-version==2.10.0
sentence-transformers==5.2.0
setuptools==80.9.0
shellingham==1.5.4
six==1.17.0
stack-data==0.6.3
starlette==0.50.0
sympy==1.14.0
tabulate==0.9.0
tenacity==9.1.2
threadpoolctl==3.6.0
tokenizers==0.22.2
tomlkit==0.13.3
torch==2.9.1
tornado==6.5.4
tqdm==4.67.1
traitlets==5.14.3
transformers==4.57.3
typer==0.21.1
typer-slim==0.21.1
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2025.3
urllib3==2.6.3
uuid_utils==0.13.0
uvicorn==0.40.0
wcwidth==0.2.14
xxhash==3.6.0
zstandard==0.25.0
python-docx
ebooklib
beautifulsoup4